Session::configure: return configured session rate or error code
Helps determine what the sample rate actually got set to since some
rounding or range bounding can occur.

+++ include/libsmu/libsmu.hpp
@@ -164,10 +164,10 @@ namespace smu {
 		/// @param sampleRate The requested sample rate for the session.
 		/// Requesting a sample rate of 0 (the default) causes the session to
 		/// use the first device's default sample rate.
-		/// @return On success, 0 is returned.
+		/// @return On success, the configured sample rate is returned.
 		/// @return On error, a negative errno code is returned.
 		/// This method may not be called while the session is active.
-		int configure(uint64_t sampleRate = 0);
+		int configure(uint32_t sampleRate = 0);
 
 		/// @brief Run the currently configured capture and wait for it to complete.
 		/// @param samples Number of samples to capture until we stop. If 0, run in continuous mode.
@@ -425,9 +425,9 @@ namespace smu {
 
 		/// @brief Configurization and initialization for device sampling.
 		/// @param sampleRate The requested sampling rate for the device.
-		/// @return On success, 0 is returned.
+		/// @return On success, the configured sample rate is returned.
 		/// @return On error, a negative errno code is returned.
-		virtual int configure(uint64_t sampleRate) = 0;
+		virtual int configure(uint32_t sampleRate) = 0;
 
 		/// @brief Turn on power supplies and clear sampling state.
 		/// @return On success, 0 is returned.
